refactor out filtering
